Issue #SB-4237feat: Remove from device pop up

+++ src/component/content-actions/content-actions.ts
@@ -1,11 +1,12 @@
 import { TelemetryGeneratorService } from './../../service/telemetry-generator.service';
 import { TranslateService } from '@ngx-translate/core';
 import { NavParams } from 'ionic-angular/navigation/nav-params';
-import { PopoverController, Events } from 'ionic-angular/index';
+ import { PopoverController, Events, AlertController } from 'ionic-angular/index';
 import { ViewController } from 'ionic-angular';
 import { Component } from '@angular/core';
 import { ContentService, AuthService, Environment, Rollup, CorrelationData, InteractType, InteractSubtype, TelemetryObject } from 'sunbird';
 import { ToastController, Platform } from 'ionic-angular';
+import { CommonUtilService } from '../../service/common-util.service';
 import { ReportIssuesComponent } from '../report-issues/report-issues';
 import { ProfileConstants } from '../../app/app.constant';
 
@@ -32,9 +33,11 @@ export class ContentActionsComponent {
     private toastCtrl: ToastController,
     public popoverCtrl: PopoverController,
     private authService: AuthService,
+    private alertctrl: AlertController,
     private events: Events,
     private translate: TranslateService,
     private platform: Platform,
+    private commonUtilService: CommonUtilService,
     private telemetryGeneratorService: TelemetryGeneratorService) {
     this.content = this.navParams.get('content');
     this.pageName = this.navParams.get('pageName');
@@ -93,7 +96,29 @@ export class ContentActionsComponent {
   close(i) {
     switch (i) {
       case 0: {
-        this.deleteContent();
+        const confirm = this.alertctrl.create({
+          title: this.commonUtilService.translateMessage('CONFIRM_DEL', this.commonUtilService.translateMessage('MSG_RESOURCE_DELETED')),
+           mode: 'wp',
+          cssClass: 'confirm-alert',
+          buttons: [
+            {
+              text: this.commonUtilService.translateMessage('CANCEL'),
+              role: 'cancel',
+              cssClass: 'alert-btn-cancel',
+              handler: () => {
+                this.viewCtrl.dismiss();
+              }
+            },
+            {
+              text: this.commonUtilService.translateMessage('REMOVE'),
+              cssClass: 'alert-btn-delete',
+              handler: () => {
+                 this.deleteContent();
+              },
+            }
+          ]
+        });
+        confirm.present();
         break;
       }
       case 1: {
